requirements.txt
amqp==2.2.2
appnope==0.1.0
awscli==1.11.177
billiard==3.5.0.3
bleach==1.5.0
boto3==1.4.7
botocore==1.7.36
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.7
decorator==4.1.2
Django==1.11.6
django-cors-headers==2.1.0
django-extensions==1.9.6
django-storages==1.6.5
djangorestframework==3.7.1
docutils==0.14
google-api-python-client==1.6.4
html5lib==0.9999999
httplib2==0.10.3
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
jmespath==0.9.3
kombu==4.1.0
Markdown==2.6.9
numpy==1.13.3
oauth2client==4.1.2
olefile==0.44
packaging==16.8
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
pip-review==1.0
prompt-toolkit==1.0.15
protobuf==3.4.0
psycopg2==2.7.3.1
ptyprocess==0.5.2
pyasn1==0.3.7
pyasn1-modules==0.1.5
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
raven==6.2.1
redis==2.10.6
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.11
selenium==3.6.0
simplegeneric==0.8.1
six==1.11.0
tensorflow==1.3.0
tensorflow-tensorboard==0.1.8
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
Werkzeug==0.12.2
